requirements.txt
streamlit
pdfplumber
sentence-transformers
faiss-cpu
numpy
scikit-learn
torch
transformers
